chore: reorganize examples (create separate directory for each example)
