+++ example/message/consumer/src/ExampleMessageConsumer.php
