+++ example/message/pacts/messageConsumer-messageProvider.json
@@ -1,0 +1,54 @@
+{
+  "consumer": {
+    "name": "messageConsumer"
+  },
+  "messages": [
+    {
+      "contents": {
+        "song": "And the wind whispers Mary"
+      },
+      "description": "footprints dressed in red",
+      "metadata": {
+        "contentType": "application/json",
+        "queue": "And the clowns have all gone to bed",
+        "routing_key": "And the clowns have all gone to bed"
+      },
+      "providerStates": [
+        {
+          "name": "You can hear happiness staggering on down the street"
+        }
+      ]
+    },
+    {
+      "contents": {
+        "text": "Hello Mary"
+      },
+      "description": "an alligator named Mary exists",
+      "metadata": {
+        "contentType": "application/json",
+        "queue": "wind cries",
+        "routing_key": "wind cries"
+      },
+      "providerStates": [
+        {
+          "name": "a message",
+          "params": {
+            "foo": "bar"
+          }
+        }
+      ]
+    }
+  ],
+  "metadata": {
+    "pactRust": {
+      "ffi": "0.4.7",
+      "models": "1.1.9"
+    },
+    "pactSpecification": {
+      "version": "3.0.0"
+    }
+  },
+  "provider": {
+    "name": "messageProvider"
+  }
+}
