--- example/pacts/README.md
@@ -1,11 +1,0 @@
-# Example Pacts
-
-The json files in this folder are explicitly here for an easy-to-read output of the test examples. These are *not* the actual test results from running all these tests of this project. By default, the pact files of this project's examples are written to example/output.  The tests themselves need to generate the appropropriate files as part of the tests.
-
-To run the tests locally, try `composer test`
-
-
-
-
-
-
