--- example/phpunit.all.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit bootstrap="../vendor/autoload.php" colors="true">
-    <testsuites>
-        <testsuite name="PhpPact Core Tests">
-            <directory>../tests</directory>
-        </testsuite>
-        <testsuite name="PhpPact Consumer Example Tests">
-            <directory>./tests/Consumer</directory>
-        </testsuite>
-        <testsuite name="PhpPact Provider Example Tests">
-            <directory>./tests/Provider</directory>
-        </testsuite>
-        <testsuite name="PhpPact Message Consumer Example Tests">
-            <directory>./tests/MessageConsumer</directory>
-        </testsuite>
-    </testsuites>
-    <php>
-        <env name="PACT_CONSUMER_NAME" value="someConsumer"/>
-        <env name="PACT_PROVIDER_NAME" value="someProvider"/>
-        <env name="PACT_OUTPUT_DIR" value=".\example\output\\"/>
-    </php>
-</phpunit>
