--- example/phpunit.consumer.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit bootstrap="../vendor/autoload.php" colors="true">
-    <testsuites>
-        <testsuite name="PhpPact Example Tests">
-            <directory>./tests/Consumer</directory>
-        </testsuite>
-    </testsuites>
-    <php>
-        <env name="PACT_CONSUMER_NAME" value="someConsumer"/>
-        <env name="PACT_PROVIDER_NAME" value="someProvider"/>
-        <env name="PACT_OUTPUT_DIR" value=".\example\output"/>
-    </php>
-</phpunit>
