--- example/phpunit.message.consumer.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit bootstrap="../vendor/autoload.php" colors="true">
-    <testsuites>
-        <testsuite name="PhpPact Example Tests">
-            <directory>./tests/MessageConsumer</directory>
-        </testsuite>
-    </testsuites>
-</phpunit>
