--- example/phpunit.provider.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit bootstrap="../vendor/autoload.php" colors="true">
-    <testsuites>
-        <testsuite name="PhpPact Provider Example Tests">
-            <directory>./tests/Provider</directory>
-        </testsuite>
-    </testsuites>
-</phpunit>
